Fixed forecast and activitylevel description

+++ Models/ForecastDay.cs
@@ -4,7 +4,7 @@ public class ForecastDay
 {
     public string Date { get; set; } = string.Empty;
     public double KpIndex { get; set; }
-    public int Probability { get; set; } //for better UX
+    public double Probability { get; set; } //for better UX
     public string ActivityLevel { get; set; } = string.Empty;
     public string IconEmoji { get; set; } = string.Empty;
